+++ s/auth/passport.ts
@@ -1,9 +1,11 @@
 
 import {deep} from "@benev/slate"
 import {Keypair} from "./keypair.js"
+import {randomId} from "./utils/random-id.js"
+import {PassportJson, KeypairJson} from "./types.js"
 import {JsonWebToken} from "./utils/json-web-token.js"
 import {randomFullName} from "../tools/random-names.js"
-import {AccessJwtPayload, PassportJson, KeypairJson} from "./types.js"
+import {LoginPayload, LoginSessionTokens, ProofPayload} from "./tokens/types.js"
 
 export class Passport {
 	constructor(
@@ -41,18 +43,42 @@ export class Passport {
 		return await Keypair.fromJson(this.keypairJson)
 	}
 
-	async signLoginToken(o: {expiry: number, audience: string, issuer: string}) {
-		const keypair = await this.getKeypair()
-		return await keypair.sign<AccessJwtPayload>({
-			sub: this.thumbprint,
-			exp: JsonWebToken.fromJsTime(o.expiry),
-			iss: o.issuer,
-			aud: o.audience,
-			data: {name: this.name, publicKey: this.keypairJson.publicKey},
-		})
-	}
+	async signLoginToken(o: {
+			expiry: number
+			issuer: string
+			audience: string
+		}): Promise<LoginSessionTokens> {
 
-	/** @deprecated renamed to `signLoginToken` */
-	signAccessToken = this.signLoginToken.bind(this)
+		const passportKeypair = await this.getKeypair()
+		const loginKeypair = await Keypair.generate()
+		const exp = JsonWebToken.fromJsTime(o.expiry)
+		const name = this.name
+		const iss = o.issuer
+		const aud = o.audience
+		const jti = await randomId()
+
+		const proofToken = await passportKeypair.sign<ProofPayload>({
+			exp,
+			iss,
+			aud,
+			jti,
+			data: {
+				loginPubkey: await loginKeypair.toPubkey().toJson(),
+				passportPubkey: await passportKeypair.toPubkey().toJson(),
+			},
+		})
+
+		const loginToken = await passportKeypair.sign<LoginPayload>({
+			sub: this.thumbprint,
+			exp,
+			jti,
+			data: {
+				name,
+				loginKeypair: await loginKeypair.toJson(),
+			},
+		})
+
+		return {proofToken, loginToken}
+	}
 }