+++ s/auth/utils/get-crypto.ts
@@ -2,8 +2,10 @@
 import {isNode} from "../../tools/is-node.js"
 
 export async function getCrypto() {
-	return isNode()
-		? (await import("crypto")).webcrypto
-		: self.crypto
+	return (
+		isNode()
+			? (await import("crypto")).webcrypto
+			: self.crypto
+	) as any as Crypto
 }