+++ s/index.ts
@@ -10,3 +10,5 @@ export * from "./federated/context.js"
 export * from "./federated/views/components.js"
 export * from "./federated/views/auth-login/element.js"
 
+export * from "./tools/storage-signal.js"
+